Move parseMessage util to global utils
