+++ src/containers/LogFace/utils.tsx
@@ -1,24 +1,6 @@
-import { replace, split, trim, trimStart } from 'lodash';
-import React from 'react';
+import { trimStart } from 'lodash';
 import queryString from 'querystring';
 import { RouteComponentProps } from 'react-router';
-
-/** convert the smsData message field from prose to more easily readable
- * point format
- *
- * @param message - the message prose.
- */
-export const parseMessage = (message: string) => {
-    const propValues = split(message, '\n');
-    const replacedEquals = propValues.map(trim).map((entry) => replace(entry, / =\s*/, ' : '));
-    return (
-        <ul>
-            {replacedEquals.map((value, index) => {
-                return <li key={`${value}-${index}`}>{value}</li>;
-            })}
-        </ul>
-    );
-};
 
 /** Get query params from URL
  *
